Add files via upload
Sometimes existing metadata standards are not appropriate for a particular dataset, or cannot be used by a particular research group for a variety of reasons. In this case it is important to look for alternatives and decide how the research group is going to document its datasets. The most simple way of recording metadata is to do it in a text file. We often refer to these files as readme files. Structuring the information is in a readme file is a good idea to make sure that the metadata will be thorough and complete, and to make it easier for members of the same team to share the data. A readme template should be tailored to the research group and to the kind of data that it is documenting to be most useful. 

A generic readme template can be found below. This template is published under a CC0 license, feel free to modify it and reuse it as you wish. This template has been designed with the goal of documenting datasets that are made publicly available in a repository, but it can be adapted to working datasets